types: move all types into a common file
