--- tpm2_pytss/ESYS_TR.py
@@ -1,49 +1,0 @@
-from ._libesys import ffi,lib
-
-class ESYS_TR(object):
-    NONE = lib.ESYS_TR_NONE
-    PASSWORD = lib.ESYS_TR_PASSWORD
-    PCR0 = lib.ESYS_TR_PCR0
-    PCR1 = lib.ESYS_TR_PCR1
-    PCR2 = lib.ESYS_TR_PCR2
-    PCR3 = lib.ESYS_TR_PCR3
-    PCR4 = lib.ESYS_TR_PCR4
-    PCR5 = lib.ESYS_TR_PCR5
-    PCR6 = lib.ESYS_TR_PCR6
-    PCR7 = lib.ESYS_TR_PCR7
-    PCR8 = lib.ESYS_TR_PCR8
-    PCR9 = lib.ESYS_TR_PCR9
-    PCR10 = lib.ESYS_TR_PCR10
-    PCR11 = lib.ESYS_TR_PCR11
-    PCR12 = lib.ESYS_TR_PCR12
-    PCR13 = lib.ESYS_TR_PCR13
-    PCR14 = lib.ESYS_TR_PCR14
-    PCR15 = lib.ESYS_TR_PCR15
-    PCR16 = lib.ESYS_TR_PCR16
-    PCR17 = lib.ESYS_TR_PCR17
-    PCR18 = lib.ESYS_TR_PCR18
-    PCR19 = lib.ESYS_TR_PCR19
-    PCR20 = lib.ESYS_TR_PCR20
-    PCR21 = lib.ESYS_TR_PCR21
-    PCR22 = lib.ESYS_TR_PCR22
-    PCR23 = lib.ESYS_TR_PCR23
-    PCR24 = lib.ESYS_TR_PCR24
-    PCR25 = lib.ESYS_TR_PCR25
-    PCR26 = lib.ESYS_TR_PCR26
-    PCR27 = lib.ESYS_TR_PCR27
-    PCR28 = lib.ESYS_TR_PCR28
-    PCR29 = lib.ESYS_TR_PCR29
-    PCR30 = lib.ESYS_TR_PCR30
-    PCR31 = lib.ESYS_TR_PCR31
-    OWNER = lib.ESYS_TR_RH_OWNER
-    NULL = lib.ESYS_TR_RH_NULL
-    LOCKOUT = lib.ESYS_TR_RH_LOCKOUT
-    ENDORSEMENT = lib.ESYS_TR_RH_ENDORSEMENT
-    PLATFORM = lib.ESYS_TR_RH_PLATFORM
-    PLATFORM_NV = lib.ESYS_TR_RH_PLATFORM_NV
-    RH_OWNER = lib.ESYS_TR_RH_OWNER
-    RH_NULL = lib.ESYS_TR_RH_NULL
-    RH_LOCKOUT = lib.ESYS_TR_RH_LOCKOUT
-    RH_ENDORSEMENT = lib.ESYS_TR_RH_ENDORSEMENT
-    RH_PLATFORM = lib.ESYS_TR_RH_PLATFORM
-    RH_PLATFORM_NV = lib.ESYS_TR_RH_PLATFORM_NV
